% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\name{diagnosis}
\alias{diagnosis}
\title{Diagnosis using hierarchical bivariate model.
 Generates a plot from CSV data.
Discussion of continuity correction:
http://handbook.cochrane.org/chapter_16/16_9_2_studies_with_zero_cell_counts.htm}
\usage{
diagnosis(content, measure, year, pmid, sortby, lefthand, righthand, type,
  cofactorlabel, topic, theme)
}
\description{
Diagnosis using hierarchical bivariate model.
 Generates a plot from CSV data.
Discussion of continuity correction:
http://handbook.cochrane.org/chapter_16/16_9_2_studies_with_zero_cell_counts.htm
}

