% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_estimators.R
\name{.om_timaeus}
\alias{.om_timaeus}
\title{Timaeus (1992) orphanhood estimation}
\usage{
.om_timaeus(sex, age, sn, mn, date)
}
\arguments{
\item{sex}{Character. \code{"Female"} or \code{"Male"}.}

\item{age}{Numeric vector. Respondent age group lower bounds.}

\item{sn}{Numeric vector. Parental survival proportions.}

\item{mn}{Numeric scalar or vector. Mean age of parents at respondent's
birth.}

\item{date}{Numeric. Survey reference date as a decimal year.}
}
\value{
A data frame with columns \code{age_input}, \code{base_age},
  \code{target_age}, \code{prob}, and \code{ref_date}.
}
\description{
Timaeus (1992) orphanhood estimation
}
\keyword{internal}
