% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schmertmann.R
\name{.qs_spline}
\alias{.qs_spline}
\title{Schmertmann quadratic spline knots and coefficients}
\usage{
.qs_spline(alpha, P, H)
}
\arguments{
\item{alpha}{Numeric. Youngest age at which fertility rises above zero.}

\item{P}{Numeric. Age at peak fertility.}

\item{H}{Numeric. Youngest age above \code{P} at which fertility falls to
half its peak.}
}
\value{
A list with \code{knots}, \code{theta}, \code{beta} and \code{W}.
}
\description{
Schmertmann quadratic spline knots and coefficients
}
\references{
Schmertmann, C. P. (2003). A system of model fertility
  schedules with graphically intuitive parameters. \emph{Demographic
  Research}, 9(5), 81-110.
}
\keyword{internal}
