% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervention.R
\name{intervention}
\alias{intervention}
\title{Perform a meta-analysis of interventions (11/30/2014 converted to Hartung and Knapp estimators for broader confidence intervals.
Generates an intervention plot from CSV data.
Alternatives:f
http://cran.r-project.org/web/packages/metafor/ (allows continuity correction)
http://cran.r-project.org/web/packages/rmeta/
Discussion of continuity correction:
http://handbook.cochrane.org/chapter_16/16_9_2_studies_with_zero_cell_counts.htm}
\usage{
intervention(content, measure, hartung, year, pmid, sortby, lefthand, righthand,
  type, independent_variable, cofactorlabel, topic, label_location, theme)
}
\description{
Perform a meta-analysis of interventions (11/30/2014 converted to Hartung and Knapp estimators for broader confidence intervals.
Generates an intervention plot from CSV data.
Alternatives:f
http://cran.r-project.org/web/packages/metafor/ (allows continuity correction)
http://cran.r-project.org/web/packages/rmeta/
Discussion of continuity correction:
http://handbook.cochrane.org/chapter_16/16_9_2_studies_with_zero_cell_counts.htm
}

