% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_estimators.R
\name{print.OrphanhoodEstimate}
\alias{print.OrphanhoodEstimate}
\title{Print method for \code{OrphanhoodEstimate} objects}
\usage{
\method{print}{OrphanhoodEstimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{OrphanhoodEstimate}.}

\item{...}{Further arguments passed to \code{\link{print.data.frame}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for \code{OrphanhoodEstimate} objects
}
