\name{qitools}
\alias{qitools}
\title{qitools}
\description{
This collection of functions implements selected tools from the following R packages at \url{https://public.opencpu.org}
- \url{http://cran.r-project.org/web/packages/qcc}
- \url{http://cran.r-project.org/web/packages/SixSigma/}
}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\seealso{
- \code{\link{charts}}.
- \code{\link{ishikawa}}.
- \code{\link{pareto}}.
- \code{\link{process map}}.
}
