% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_estimators.R
\name{summary.OrphanhoodEstimate}
\alias{summary.OrphanhoodEstimate}
\title{Summary method for \code{OrphanhoodEstimate} objects}
\usage{
\method{summary}{OrphanhoodEstimate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{OrphanhoodEstimate}.}

\item{...}{Further arguments (currently unused).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for \code{OrphanhoodEstimate} objects
}
