% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{app_HOME}
\alias{app_HOME}
\title{Launch the HOME Shiny App}
\usage{
app_HOME()
}
\value{
A Shiny App object.
}
\description{
Opens the interactive dashboard for Harmonized Orphanhood Mortality Estimation.
}
