% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_diagnostics.R
\name{om_plot_linearity}
\alias{om_plot_linearity}
\title{Diagnostic Plot: Internal Consistency Check}
\usage{
om_plot_linearity(object)
}
\arguments{
\item{object}{An object of class \code{OrphanhoodEstimate}.}
}
\value{
A ggplot object.
}
\description{
Displays the estimated alpha parameter across respondent ages.
}
