% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_diagnostics.R
\name{plot.OrphanhoodSensitivity}
\alias{plot.OrphanhoodSensitivity}
\title{Plot Sensitivity to Mean Age of Childbearing}
\usage{
\method{plot}{OrphanhoodSensitivity}(x, index = "e30", ...)
}
\arguments{
\item{x}{An object of class \code{OrphanhoodSensitivity}.}

\item{index}{Character. Mortality index to plot ("e30", "45q15", or "30q30").}

\item{...}{Unused.}
}
\description{
Plot Sensitivity to Mean Age of Childbearing
}
