% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_diagnostics.R
\name{plot.OrphanhoodSensitivityFamily}
\alias{plot.OrphanhoodSensitivityFamily}
\title{Plot Model Life Table Family Sensitivity}
\usage{
\method{plot}{OrphanhoodSensitivityFamily}(x, index = "e30", ...)
}
\arguments{
\item{x}{An object of class \code{OrphanhoodSensitivityFamily}.}

\item{index}{Character. Mortality index to plot.}

\item{...}{Unused.}
}
\description{
Plot Model Life Table Family Sensitivity
}
