\name{qitools}
\alias{qitools}
\title{qitools}
\description{
This collection of functions implements selected tools from 
\url{http://cran.r-project.org/web/packages/qcc} at 
\url{https://public.opencpu.org}
}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}, Scott Chamberlain, Jeroen Ooms
}
\seealso{
\code{\link{ishikawa}}.
\code{\link{make_charts}}.
\code{\link{pareto}}.
}
