\name{qitools}
\alias{qitools}
\title{qitools}
\description{
This collection of functions implements selected tools from the following R packages at \url{https://public.opencpu.org}

\itemize{
\item \url{http://cran.r-project.org/web/packages/qcc}
\item \url{http://cran.r-project.org/web/packages/SixSigma}
}

}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\seealso{
\itemize{
\item \code{\link{charts}}.
\item \code{\link{ishikawa}}.
\item \code{\link{pareto}}.
\item \code{\link{processmap}}.
}
}

http://cran.r-project.org/doc/manuals/R-exts.html
